refactor: Header 컴포넌트 customHook 분리
